feat(contrib): add Nix flake using Zig
As `neovim-nightly-overlay` seems to be broken due to unresolved memory leak issues (see #26952), using Zig as a fallback to compile Neovim and run the test suites is our best bet not to pull a bunch of dependencies into the Nix shell. Tested locally on NixOS 25.11 x86_64, everything seems to be working fine.

The flake also exposes an executable `nvim-dev` shell script to run the produced binary from `zig-out/bin` with the proper `VIMRUNTIME` environment variable and accepts passing arguments to the `nvim` binary.